fix(windows): canonicalize rooted-but-driveless paths via drive-root probe
assertExternalDirectory and read both call AppFileSystem.normalizePath, but
on Windows that fell through to path.resolve which silently bound rooted-but-
driveless paths (/users/runner/...) to cwd's drive. When the file lived on
a different drive (typical on GitHub runners: temp on C: vs workspace on D:)
the resulting permission glob used the wrong drive.

Probe each known drive root for an existing path before falling back to
win32.resolve. Move the helper into core/filesystem and delegate from
opencode/util/filesystem so both paths share a single implementation.

Closes #427
